Design

+++ src/main/java/com/joutvhu/expansy/match/filter/TrackPoint.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class TrackPoint {
-    private long index;
+    private int index;
     private String value;
 }
